feat(functionality): fix some bugs and optimize chat list and information display
- modify chat list sorting logic, filter out records with order as null
- add message notification TODO
- optimize user information display
- update current user ID acquisition logic
- refactor some code to improve maintainability

feat(functionality): 修复一些bug和优化聊天列表和信息展示

- 修改聊天列表排序逻辑，过滤掉 order 为 null 的记录
- 添加消息通知的TODO
- 优化用户信息展示
- 更新当前用户 ID 获取逻辑
- 重构部分代码以提高可维护性

+++ app/src/main/java/com/gohj99/telewatch/ui/ChatInfoScreen.kt
@@ -67,7 +67,7 @@ fun SplashChatInfoScreen(
 chatObject: TdApi.Chat,
 subtitle: String,
 info: String,
-deleteChat: () -> Unit
+deleteChat: (() -> Unit)? = null
 ) {
     val gson = Gson()
     var messageJson = ""
@@ -244,26 +244,28 @@ deleteChat: () -> Unit
                     }
                 }
             }
-            item {
-                // 删除聊天
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .fillMaxHeight()
-                        .background(
-                            Color.Black.copy(alpha = 0.8f),
-                            shape = RoundedCornerShape(8.dp)
+            if (deleteChat != null) {
+                item {
+                    // 删除聊天
+                    Box(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .fillMaxHeight()
+                            .background(
+                                Color.Black.copy(alpha = 0.8f),
+                                shape = RoundedCornerShape(8.dp)
+                            )
+                            .padding(7.dp),
+                        contentAlignment = Alignment.Center
+                    ) {
+                        CustomButton(
+                            text = stringResource(R.string.Delete_Chat),
+                            color = Color(0xFFF44336),
+                            onClick = {
+                                deleteChat()
+                            }
                         )
-                        .padding(7.dp),
-                    contentAlignment = Alignment.Center
-                ) {
-                    CustomButton(
-                        text = stringResource(R.string.Delete_Chat),
-                        color = Color(0xFFF44336),
-                        onClick = {
-                            deleteChat()
-                        }
-                    )
+                    }
                 }
             }
             item {
